Changed List<Integer> to List<Float> per code from Tom. Updated list of difficulties to match the correct values.

+++ src/main/java/com/kyrlach/issuetracker/issue/IssueRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface IssueRepository extends CrudRepository<Issue, Long> {
 		
-	List<Issue> findByTitleLikeAndDescriptionLikeAndCategoryInAndDifficultyIn(String title, String description, List<String> categories, List<Integer> difficulties);
+	List<Issue> findByTitleLikeAndDescriptionLikeAndCategoryInAndDifficultyIn(String title, String description, List<String> categories, List<Float> difficulties);
 
 }
